commit lesson 1_2
